test_project/Personal-Django-main/requirements.txt
# Core Django packages
Django>=4.2.0,<5.0
djangorestframework==3.15.2
drf-yasg==1.21.11
# HTTP and networking
requests>=2.25.1
# Database connectors
psycopg2-binary==2.9.11
pymysql==1.1.2
# Utilities
python-dotenv==1.0.1
# Production server
gunicorn==23.0.0
# CORS handling
django-cors-headers==4.6.0
# Static file serving
whitenoise==6.8.2
# JWT Authentication
djangorestframework-simplejwt==5.3.1
# TOTP / Google Authenticator
pyotp==2.9.0